requirements2.txt
anyio==2.2.0
argon2-cffi==20.1.0
async-generator==1.10
attrs==19.3.0
Automat==0.8.0
Babel==2.9.0
backcall==0.2.0
bleach==3.3.0
blinker==1.4
certifi==2019.11.28
cffi==1.14.5
chardet==3.0.4
Click==7.0
cloud-init==20.4
colorama==0.4.3
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cryptography==2.8
dbus-python==1.2.16
decorator==4.4.2
defusedxml==0.7.1
distro==1.4.0
distro-info===0.23ubuntu1
entrypoints==0.3
httplib2==0.14.0
hyperlink==19.0.0
idna==2.8
importlib-metadata==1.5.0
incremental==16.10.1
ipykernel==5.5.0
ipython==7.21.0
ipython-genutils==0.2.0
jedi==0.18.0
Jinja2==2.10.1
json5==0.9.5
jsonpatch==1.22
jsonpointer==2.0
jsonschema==3.2.0
jupyter-client==6.1.12
jupyter-core==4.7.1
jupyter-packaging==0.7.12
jupyter-server==1.4.1
jupyterlab==3.0.12
jupyterlab-pygments==0.1.2
jupyterlab-server==2.3.0
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
MarkupSafe==1.1.0
mistune==0.8.4
more-itertools==4.2.0
nbclassic==0.2.6
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.2
nest-asyncio==1.5.1
netifaces==0.10.4
notebook==6.2.0
numpy==1.20.1
oauthlib==3.1.0
packaging==20.9
pandocfilters==1.4.3
parso==0.8.1
pexpect==4.6.0
pickleshare==0.7.5
prometheus-client==0.9.0
prompt-toolkit==3.0.17
ptyprocess==0.7.0
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycparser==2.20
Pygments==2.8.1
PyGObject==3.36.0
PyHamcrest==1.9.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyparsing==2.4.7
pyrsistent==0.15.5
pyserial==3.4
python-apt==2.0.0+ubuntu0.20.4.4
python-dateutil==2.8.1
python-debian===0.1.36ubuntu1
pytz==2021.1
PyYAML==5.3.1
pyzmq==22.0.3
requests==2.22.0
requests-unixsocket==0.2.0
scipy==1.6.1
SecretStorage==2.3.1
Send2Trash==1.5.0
service-identity==18.1.0
SimpleITK-SimpleElastix @ file:///home/rgologorsky/DeepPit/SimpleITK_SimpleElastix-2.0.0rc2.dev909-cp38-cp38-manylinux_2_24_x86_64.whl
simplejson==3.16.0
six==1.14.0
sniffio==1.2.0
sos==4.0
ssh-import-id==5.10
systemd-python==234
terminado==0.9.3
testpath==0.4.4
tornado==6.1
traitlets==5.0.5
Twisted==18.9.0
ubuntu-advantage-tools==20.3
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
wadllib==1.3.3
wcwidth==0.2.5
webencodings==0.5.1
zipp==1.0.0
zope.interface==4.7.1
